Changes:
Change to key handling

Changes to PasswordManager class as it now prompts the user to set path to set and store the AES key.

Changes the EncryptionUtil class to handle AES key from the location that the user set it to, and generate it to that location if it doesn't exist.